MOVE #48 logic code into own directory
