+++ src/logic/IPlayer.ts
