+++ src/logic/IWareForCity.ts
