+++ src/logic/i-inventory.ts
