+++ src/logic/inventory.test.ts
