+++ src/logic/wareType.ts
