+++ src/scenes/data-registry/getCities.ts
@@ -1,5 +1,5 @@
 import { Data } from "phaser";
-import { ICity } from "../ICity";
+import { ICity } from "../../logic/ICity";
 import { KEYS } from "../keys";
 
 export const getCities = (scene: { registry: Data.DataManager }): ICity[] =>
